MicroService/requirements.txt
flask>=2.0.0
influxdb-client>=1.21.0
kafka-python>=2.0.2
tensorflow>=2.5.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
psutil>=5.8.0
